requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flask-form==1.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.1
python-dotenv==1.0.0
scikit-learn==1.3.2
scipy==1.11.3
sklearn==0.0.post11
threadpoolctl==3.2.0
Werkzeug==3.0.1
WTForms==3.1.1